refactor(adapters): extract vulnerability renderers to vuln_render.rs [#305-5/7]
Extract 5 vulnerability rendering methods from markdown_formatter/mod.rs
into a dedicated vuln_render.rs submodule as pub(super) free functions:
- render_vulnerabilities
- render_vulnerability_summary
- render_actionable_vulnerabilities
- render_informational_vulnerabilities
- render_vulnerability_row

Move 5 corresponding unit tests to vuln_render.rs.
Update mod.rs to delegate to vuln_render:: functions.

Closes #327
